Move extensions & clean
